Refactored source code
