--- thot-app/src/main/java/org/cc/app/INameValidator.java
@@ -1,8 +1,0 @@
-package org.cc.app;
-
-/**
- * Created by adi on 8/21/14.
- */
-public interface INameValidator {
-    Boolean isValid(String speaker);
-}
