+++ thot-app/src/main/java/org/cc/thot/util/XlsSessionReader.java
@@ -1,4 +1,4 @@
-package org.cc.app;
+package org.cc.thot.util;
 
 import java.io.BufferedInputStream;
 import java.io.FileInputStream;
