+++ thot-app/src/main/java/org/cc/thot/web/controller/CommentsController.java
@@ -6,11 +6,11 @@ import java.net.URLDecoder;
 import javax.servlet.http.HttpServletRequest;
 
 import org.apache.log4j.Logger;
-import org.cc.app.TwitterLinkCreator;
 import org.cc.thot.dao.CommentDao;
 import org.cc.thot.dao.LinkDao;
 import org.cc.thot.dao.SessionDao;
 import org.cc.thot.domain.Session;
+import org.cc.thot.util.TwitterLinkCreator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
