+++ thot-app/src/main/java/org/cc/thot/web/controller/ProposeSessionController.java
@@ -8,7 +8,6 @@ import org.cc.thot.dao.TimeslotDao;
 import org.cc.thot.domain.Session;
 import org.cc.thot.domain.SessionType;
 import org.cc.thot.model.OpenSpaceFormData;
-import org.cc.thot.web.validate.OpenSpaceValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -24,11 +23,13 @@ public class ProposeSessionController {
 
     @Autowired
     private SessionDao sessionDao;
+    
     @Autowired
     private TimeslotDao timeslotDao;
+    
     @Autowired
     private LocationDao locationDao;
-
+    
     public ProposeSessionController() {
 	super();
     }
@@ -50,10 +51,6 @@ public class ProposeSessionController {
     public ModelAndView processSubmit(HttpServletRequest request, ModelMap modelMap, @ModelAttribute("sessionDataFormData") OpenSpaceFormData formData, BindingResult result) {
 
 	modelMap.put("sessionDataFormData", formData);
-
-	// do validation
-	OpenSpaceValidator.nameValidator = new NameValidator();
-	OpenSpaceValidator.validate(formData, result);
 
 	if (result.hasErrors()) {
 	    System.out.println("Validation errors ocurred.");
