+++ thot-app/src/main/java/org/cc/thot/web/validate/NameValidator.java
@@ -1,0 +1,10 @@
+package org.cc.thot.web.validate;
+
+/**
+ * Created by adi on 8/21/14.
+ */
+public interface NameValidator {
+
+    public boolean isValid(String speaker);
+
+}
