+++ thot-app/src/main/java/org/cc/thot/web/validate/impl/Time24HoursValidator.java
@@ -1,4 +1,4 @@
-package org.cc.thot.web.validate;
+package org.cc.thot.web.validate.impl;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
