+++ thot-app/src/main/webapp/WEB-INF/jsp/footer.html
@@ -4,9 +4,9 @@
         <spring:message code="footer.about"
 			text="This is a community app developed by participants of " />
 		<a target="_blank" href="http://www.javaland.eu/"><spring:message
-				code="footer.conferencename" text="ALE 2014" /></a>
+				code="footer.conferencename" text="JavaLand 2015" /></a>
 		<spring:message code="footer.conferencelocation"
-			text="(Krakow, Poland)" />
+			text="(Cologne, Germany)" />
 		<br />
 		<spring:message code="footer.sposoredby"
 			text="Hosting is sponsord by " />
@@ -15,9 +15,9 @@
       </div>
     </footer>
 
-<script src="<%= request.getContextPath() %>/assets/js/jquery.js"></script>
-<script src="<%= request.getContextPath() %>/assets/js/bootstrap.js"></script>
-<script src="<%= request.getContextPath() %>/assets/js/thot.js"></script>
+<script src="<%= request.getContextPath() %>/js/jquery.js"></script>
+<script src="<%= request.getContextPath() %>/js/bootstrap.js"></script>
+<script src="<%= request.getContextPath() %>/js/thot.js"></script>
 
 </body>
 </html>
