+++ thot-app/src/main/webapp/WEB-INF/jsp/searchSessions.jsp
@@ -1,7 +1,7 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <%@ taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@ taglib uri="http://www.springframework.org/tags" prefix="spring"%>
-<%@ page import="org.cc.app.TwitterLinkCreator"%>
+<%@ page import="org.cc.thot.util.TwitterLinkCreator"%>
 <%@ page session="false"%>
 
 <%@ include file="header.jsp"%>
