+++ thot-app/src/test/java/org/cc/thot/util/XlsSessionReaderTest.java
@@ -1,4 +1,4 @@
-package org.cc.app;
+package org.cc.thot.util;
 
 import org.junit.Test;
