+++ thot-app/src/test/java/org/cc/thot/web/controller/NameValidatorTests.java
@@ -3,7 +3,7 @@ package org.cc.thot.web.controller;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import org.cc.thot.web.controller.NameValidator;
+import org.cc.thot.web.validate.impl.NameValidatorImpl;
 import org.junit.Test;
 
 /**
@@ -13,7 +13,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameValidWhenContainsNumbers() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@243";
 
 	assertTrue(nameValidator.isValid(emptySpeaker));
@@ -21,7 +21,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameValidWhenContainsLowerCaseLetters() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@validname";
 
 	assertTrue(nameValidator.isValid(emptySpeaker));
@@ -29,7 +29,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameValidWhenContainsUpperCaseLetters() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@ValidName";
 
 	assertTrue(nameValidator.isValid(emptySpeaker));
@@ -37,7 +37,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameValidWhenContainsUnderscore() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@some_name";
 
 	assertTrue(nameValidator.isValid(emptySpeaker));
@@ -45,7 +45,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameNotValidWhenContainsDash() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@some-name";
 
 	assertFalse(nameValidator.isValid(emptySpeaker));
@@ -53,7 +53,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameIsNotValidWhenEmpty() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "";
 
 	assertFalse(nameValidator.isValid(emptySpeaker));
@@ -61,7 +61,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameIsNotValidWhenNotStartingWithAt() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@";
 
 	assertFalse(nameValidator.isValid(emptySpeaker));
@@ -69,7 +69,7 @@ public class NameValidatorTests {
 
     @Test
     public void nameIsNotValidWhenContainsSpace() {
-	NameValidator nameValidator = new NameValidator();
+	NameValidatorImpl nameValidator = new NameValidatorImpl();
 	String emptySpeaker = "@some name";
 
 	assertFalse(nameValidator.isValid(emptySpeaker));
