--- thot-app/src/test/java/org/cc/thot/web/validate/OpenSpaceFormDataBuilder.java
@@ -1,36 +1,0 @@
-package org.cc.thot.web.validate;
-
-import org.cc.thot.model.OpenSpaceFormData;
-
-/**
- * Created by adi on 8/21/14.
- */
-public class OpenSpaceFormDataBuilder {
-    private OpenSpaceFormData openSpaceFormData;
-    private String title = "Correct title";
-    private String speaker = "@someValidName";
-
-    public OpenSpaceFormDataBuilder valid() {
-	openSpaceFormData = new OpenSpaceFormData();
-	openSpaceFormData.setSpeaker(speaker);
-	openSpaceFormData.setTitle(title);
-	return this;
-    }
-
-    public OpenSpaceFormDataBuilder withTitle(String title) {
-	this.title = title;
-	return this;
-    }
-
-    public OpenSpaceFormDataBuilder withSpeaker(String speaker) {
-	this.speaker = speaker;
-	return this;
-    }
-
-    public OpenSpaceFormData build() {
-	openSpaceFormData = new OpenSpaceFormData();
-	openSpaceFormData.setSpeaker(speaker);
-	openSpaceFormData.setTitle(title);
-	return openSpaceFormData;
-    }
-}
